requirements.txt
Django==4.2
djangorestframework==3.14.0
PyJWT==2.8.0
python-dotenv==1.0.0
django-ratelimit==3.0.1